reduce deployment size to 32MB

+++ core/entrypoint.sh
@@ -1,10 +1,10 @@
-#!/usr/bin/env bash
+#!/bin/sh
+
 set -xeuo
 
 set -a
 . /opt/strimserver/config/strimserver.env
 set +a
-
 
 # Optional: load from Docker secret file if env var not already set
 if [ -z "${SRT_PUBLISH_PASSPHRASE:-}" ] && [ -f /run/secrets/srt-passphrase ]; then
@@ -13,15 +13,21 @@ if [ -z "${SRT_PUBLISH_PASSPHRASE:-}" ] && [ -f /run/secrets/srt-passphrase ]; t
 fi
 
 # Fail fast if missing
+: "${NVIDIA_DRIVER_VERSION_MAJOR:?NVIDIA_DRIVER_VERSION_MAJOR is required}"
 : "${NORMALIZED_MPEGTS_SOCKET:?NORMALIZED_MPEGTS_SOCKET is required}"
 
+cd /usr/lib64 \
+   && ln -sf libcuda.so.$NVIDIA_DRIVER_VERSION_MAJOR* libcuda.so.1 \
+   && ln -sf libnvcuvid.so.$NVIDIA_DRIVER_VERSION_MAJOR* libnvcuvid.so.1 \
+   && ln -sf libnvidia-ptxjitcompiler.so.$NVIDIA_DRIVER_VERSION_MAJOR* libnvidia-ptxjitcompiler.so.1 \
+   && ln -sf libnvidia-encode.so.$NVIDIA_DRIVER_VERSION_MAJOR* libnvidia-encode.so.1 \
+   && cd
 
 # Render final config
-mkdir -p /opt/strimserver/runtime
-envsubst < /opt/strimserver/config/mediamtx.yaml.template > /opt/strimserver/runtime/mediamtx.yaml
+envsubst < /opt/strimserver/config/mediamtx.yaml.template > /opt/strimserver/config/mediamtx.yaml
 
 
 rm -f "$NORMALIZED_MPEGTS_SOCKET"
 
-exec /usr/bin/nice -n -10 /usr/local/bin/mediamtx /opt/strimserver/runtime/mediamtx.yaml
+exec /usr/bin/nice -n -10 /usr/local/bin/mediamtx /opt/strimserver/config/mediamtx.yaml
